requirements.txt
alembic==1.0.11
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dotenv==0.10.3
six==1.12.0
WebOb==1.8.5
Werkzeug==0.15.4
psycopg2-binary==2.8.3
SQLAlchemy==1.3.5
Flask-Sockets==0.2.1
redis==3.3.4
